Apply some refactors to have a more readable code
