--- src/main/java/org/eyeseetea/sdk/common/EyeSeeTeaSdkInstance.java
@@ -1,25 +1,0 @@
-package org.eyeseetea.sdk.common;
-
-public class EyeSeeTeaSdkInstance {
-
-    private static EyeSeeTeaSdkInstance mInstance;
-    private Transaltor mTransaltor;
-
-    private EyeSeeTeaSdkInstance() {
-    }
-
-    public static EyeSeeTeaSdkInstance getInstance() {
-        if (mInstance == null) {
-            mInstance = new EyeSeeTeaSdkInstance();
-        }
-        return mInstance;
-    }
-
-    public Transaltor getTranslator() {
-        return mTransaltor;
-    }
-
-    public void initTranslator(Transaltor transaltor) {
-        mTransaltor = transaltor;
-    }
-}
